Honor remaining configurable limits and stabilize deadline tests
Use configured MCP label and MIME bounds in final projections and retain
truncation metadata. Widen the MCP line counter to avoid sentinel overflow.
Apply the Anthropic discovery limit consistently to exact-model overrides.
Retain only complete UTF-8 scalars when capturing truncated Git blobs.

Inject TimeProvider into semantic query deadlines with scoped timer and
linked-token ownership. Test an expired deadline deterministically instead
of relying on a 1 ms operating-system timer race. Retry transient Windows
build-fixture directory cleanup only after process-exit assertions.

Add regression coverage for the affected configuration, counter, Unicode,
and timeout paths. Preserve existing defaults and update capture-unit docs.

+++ src/Threadsmith.Mcp/McpBoundedLineReadStream.cs
@@ -8,7 +8,7 @@ internal sealed class McpBoundedLineReadStream : Stream
 
     private readonly int _maximumLineBytes;
     private readonly Stream _inner;
-    private int _currentLineBytes;
+    private long _currentLineBytes;
 
     /// <summary>Initializes a new instance of the <see cref="McpBoundedLineReadStream"/> class.</summary>
     internal McpBoundedLineReadStream(Stream inner, int maximumLineBytes = MaximumLineBytes)
